requirements.txt
fastapi
uvicorn
mysql-connector-python
beautifulsoup4
python-dotenv
requests
groq